fix: move getProjectRoot, add scope tests, add docs
